backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
numpy==1.26.2
librosa==0.10.2
tensorflow==2.15.0
tensorflow-hub==0.16.1
pymongo==4.6.0
redis==5.0.1
python-dotenv==1.0.0
pydub==0.25.1
soundfile==0.12.1
scipy==1.11.4
bcrypt==4.1.2
PyJWT==2.8.0

